+++ glm_asr_triton_template_updated/attention.py
@@ -56,7 +56,7 @@ def attention_scores_kernel(
         q_ptr + pid_bh * stride_q0 + pid_q * stride_q1 + offs_d * stride_q2,
         mask=offs_d < head_dim,
         other=0.0,
-    ).to(tl.float32)
+    )
     k = tl.load(
         k_ptr
         + pid_bh * stride_k0
@@ -64,7 +64,7 @@ def attention_scores_kernel(
         + offs_d[None, :] * stride_k2,
         mask=(offs_k[:, None] < seq_k) & (offs_d[None, :] < head_dim),
         other=0.0,
-    ).to(tl.float32)
+    )
     scores = tl.sum(k * q[None, :], axis=1) * scale
     tl.store(
         scores_ptr
@@ -131,7 +131,7 @@ def attention_output_kernel(
         + offs_k * stride_w2,
         mask=offs_k < seq_k,
         other=0.0,
-    ).to(tl.float32)
+    )
     v = tl.load(
         v_ptr
         + pid_bh * stride_v0
@@ -139,7 +139,7 @@ def attention_output_kernel(
         + offs_d[None, :] * stride_v2,
         mask=(offs_k[:, None] < seq_k) & (offs_d[None, :] < head_dim),
         other=0.0,
-    ).to(tl.float32)
+    )
     out = tl.sum(v * w[:, None], axis=0)
     tl.store(
         output_ptr
